handel: Move the network test as unit test
Move the `it_buffers_messages` handel network test from a file in
the `test` directory to the same file the code is to make it an
unit test instead of an integration test.
